Moved Grabber interface out of Robot class.
